Refactor: model 디렉토리 이름 => store로 변경
헷갈려서 통합함
